Rewritten the decode encode test using the new decorator
Also limited it to the first 100 chunks it finds to limit the amount of time it takes
